refactor: Move code for legacy mod install reading to separate file (#416)
Moves some logic related to reading legacy installed mods to new file in order to clean up code before adding new logic.
